Refactor email service
