--- src/Core/MoneyManager.Application/Contracts/Services/IEmailSender.cs
@@ -1,8 +1,0 @@
-﻿namespace MoneyManager.Application.Contracts.Services
-{
-    public interface IEmailSender
-    {
-        Task SendForgotPasswordEmailAsync(string urlToResetPassword, string recipient);
-        Task SendChangeEmialEmailAsync(string ulrToChangeEmail, string recipient);
-    }
-}
