+++ src/Core/MoneyManager.Application/Contracts/Services/IEmailService.cs
@@ -1,0 +1,8 @@
+﻿namespace MoneyManager.Application.Contracts.Services
+{
+    public interface IEmailService
+    {
+        Task SendForgotPasswordEmailAsync(string recipient, string token);
+        Task SendChangeEmailEmailAsync(string recipient, string keyConfirmingEmailChange);
+    }
+}
