+++ src/Infrastructure/MoneyManager.Infractructure/Services/EmailService/EmailSender/IEmailSender.cs
@@ -1,0 +1,7 @@
+﻿namespace MoneyManager.Infractructure.Services.EmailService.EmailSender
+{
+    public interface IEmailSender
+    {
+        Task SendEmailWithHtmlBody(string recipient, string subject, string body);
+    }
+}
